-Hyperlinks and "table of contents" added to FAQ
-Starting and stopping translations redesigned to give the user more
precise control by having a master "Allow translations" checkbox in
Preferences and a Stop and Start button in the Translations tab

-Progress window slightly redesigned to tell the user it may take a
while depending on the number of documents and their computer

-"Show/hide warnings" finally fully implemented for loading/removing
documents in the documents tab or the PPSP and GUIMain frames.

-JFileChoosers updated for PPSP and GUIMain to remember directories to
make constructing problem sets easier

-Autoscroll turned off for translations

-User can now delete any of their documents after processing without
breaking Anonymouth

-"Save" no longer immediately writes over original document, it now
operates as "Save As" if the user hasn't yet saved their progress, and
then acts as "Save" to that new file every subsequent use.