It is done! No known bugs at this time (besides the Splash Screen text not updating, not a big deal). Note also that the Advanced Set up window UI has not yet been updated, and until then the Advanced button does nothing.
-ExtFilter moved out of being an odd inner class in one of the classes
in gooie and moved as it's own class to helpers
-Lots of cleaning up and old/unused variables deleted
-Complete GUI and Backend redesign of the Pre-processing tab, now it's
own window as a set up wizard
-Modified the PreProcessingSettingsFrame class to be the "advanced"
button of the Pre-processing wizard
-Splash screen added to display text progress as Anonymouth cold loads
(takes a while sometimes, we want to show the user things are being
done)
-ThePresident cleaned up and organized, anything that isn't a constant
isn't all caps
-Created ANONConstants which acts much like JSANConstants, cleans up
ThePresident and consolidates all application-wide constants
-ImageLoader class created in helpers to help ease the annoyance of
having to write a new try-catch and specific load code for the kind of
image you want
-Logger can now buffer printed output if file isn't created yet and then
flush as soon as the file is created (so we can get the start up logs
even though the session name wasn't yet created, and thus the logger
file)
-Anonymouth now used the system's default file chooser instead of java's
ugly and kinda shitty version
-The file extension filter now actually works
-Numerous bug fixes, usability tweaks, and fine details
-The user can now delete nodes from the train tree or sample list by
using the delete key
-Now, if the user tries to add just files instead of a whole directory,
Anonymouth will use the name of the parent directory the files are
residing in (which most of the time is the name the user wants and looks
better that "no author entered")
-The user can now rename authors if they don't want the parent
directory's name being used
-The session name prompt is now in it's own window separate from the
starting window
-The session name window can now easily be added/removed from Anonymouth
by it's corresponding flag in ANONConstants
-Documentation added to all new classes and methods.
ColumnsAutoSizer now moved out of it's home in GUIMain (wasn't even used
in there) and now rests in helpers.
-Loading saved problem sets is no longer a passive process, it now
actively checks each file to make sure they're not empty/make sure they
exist/make sure we have read permissions, and then checks to see if with
these omissions (if any) we still have a satisfactory document set. If
we do, the user may begin. If not, they can go in to the wizard and add
back in the few files they need to begin.