-The PreProcessingSettingsFrame "Features" tab has been completely revised, stripped down for clarity, and cleaned up in PreProcessAdvancedWindow.
-The bottom section of the old "Features" tab has been put hidden away
in it's own window so those that want to dig in that deep may, but it
won't bother people that don't want to be confused or bothered by it.

-The Feature information viewer now clears and handles data properly
(Was not clearing and displaying data correctly before).

-The PreProcessingSettingsFrame "Classifier" tab has been completely
revised, stripped down for clarity, and cleaned up in
PreProcessAdvancedWindow.

-The PreProcessAdvancedWindow is now completely online and functional.

-The Classifier arguments pane is now in an editor window, where it will
only allow you to use acceptable parameters (it checks as you type as
enables or disables the "Ok" button accordingly).

-ScrollToTop helper class added to helpers which gives us the ability to
scroll a pane to the top after just updating it's contents (the build in
calls to do this do not work when text has been updated recently unless
you utilize an invokeLater call to ensure the text has been updated
completely before scrolling).

-Numerous other bug fixes and improvements regarding the PreProcessing
side of Anonymouth.