-JSANConstants consolidated into ANONConstants, constant names simplified.
-Word Suggestions tab edited slightly, now with redesigned color borders
and "Highlight All" option for the Words to Remove

-The user can now hide or show nearly any part of Anonymouth they want,
depending on what they like to use or need/don't need at the time (this
also allows by design for a "distraction free" mode for proofreading)

-Functionally translations are completely revamped. Instead of
translating the entire doc (which is kind of ridiculous since it takes
so much time, slows down the editor, and is extremely prone to breaking)
we have a single button that allows the user to translate any number of
single sentences. Far more stable, easier to use, and useful.

-Visually translations are completely revamped. The progress label/bar,
start button, and all that other stuff that was scattered everywhere is
now consolidated on the top part of the tab and switches from the start
button to the progress bar to hiding itself, depending on what
translation action is currently being done on the current sentence.

-Translator "swap" arrow redesigned in Java 2D Graphics and now changes
color depending on it's anonymity index to inform the users that they
should try to use green over red translations if possible.

-TaggedSentences are now cleaned up and deleted when being removed
(before the taggedSentences would just sort of stick around, not sure if
garbage collection would ever take care of them or not, staying on the
safe side).

-Translation swap arrow animation triggers tweaked.

-Undo/Redo control now works completely with the redesigned translations
backend.