-Translations help button now displays a help popup.
-All focus and focus traversal disabled in GUIMain and Preferences,
doesn't make any sense to have it there.

-DisableFocus helper added that traverses recursively through component
hierarchies to disable focus on all components (instead of having to do
component.setFocusable(false) manually for every component, which is
currently the only way to do so in Java.

-The "X" button for the progress window no longer just closes the window
and continues processing in the background without telling the user or
given them a way to stop it. Now a popup shows up telling them that
their document is still processing and whether or not they would like to
cancel or quit.