My last major Anonymouth commit. It's been a blast Andrew, thanks for everything.
-SentenceMaker has been completely redesigned to be more efficient and
easier to read.

-New "TextWrapper" implementation in place which allows for vastly
improved quotations, parenthesis, brackets, citations, etc. insertion,
deletion, and manipulation

-Big bug in editor discovered and patched which would split the sentence
up funny when the user adds to the very end of an existing sentence
(before the EOS character).

-Fixed unicode formatting error where it would not properly substitute
or remove all unicode characters, thus breaking the editor for text
formatted this way.

-Fixed an issue where white space would not be trimmed from the
beginning or end of a document, thus throwing the length calculation off
(since it wouldn't be copied over to the text in the editor later down
the line) and thus removing the ability to type at the end of a document

-Fixed a bug where the Train Docs Tree in PreProcessDriver would not
have it's model updated correctly (which would led to the table not
updating on the user's end).

-Fixed a bug with the translations swapper where it wouldn't update all
the backend properly and break the SpecialCharTracker.

-Minor performance improvements in the editor

-Minor highlighting tweak which ensures the to remove auto highlights
don't extend past the sentence.