A couple more bug fixes for the cluster window. First, fixed a bug in Anonymouth's version of JStylo that "culled" all features in the testing document that did not also appear in the first doucment of the training set, but then added these features to the instances anyway and incorrectly assigned them a value of 0. I removed the line that culled the features, as it looks like feautres are only added if they appear in the training set anyway, making the culling unnecessary. This bug is not present in the current standalone version of JStylo.
Also fixed a bug where feature values in the cluster window would not
update in many cases. Originally, we relied on TaggedDocument's
updateReferences method, which did not update document level features
such as character count and did not update features if the taggedDoc was
replaced by one from the undo/redo stacks. Now, every time the document
is reprocessed feature values are updated using the current instance
from DocumentMagician, so they should always be accurate to the last
reprocess.