Major improvements to this branch's processing time. The anonymity bar, word suggestions, and translations will now all use the same InstanceConstructor that DocumentMagician uses for initial feature extraction, meaning that they can extract features from the testing document without re-extracting the training data first. FeatureSwapper will now load the pre-trained classifier that DocumentMagician saves, so now it doesn't have to extract anything or build a new classifier.
Combined, these changes cut the jstyloUpdate branch's processing time in
half for me. Initial processing for one of the sample document sets used
to take more than 3 minutes on my computer; the same set now takes
around 90 seconds. Note however that the dictionary branch took a little
over 100 seconds, and that's with the unnecessary feature extraction for
the word suggestions and anonymity bar, so there's probably still room
for improvement.