Updated the SampleDocs panel in PreProcessWindow.
Made it so that users can select directories, which recursively traverses
any subdirectories, adding any TXT files inside them.

Fixed the UI so that the components are not cut off (The JList was cut
off, making the scrollbar not appear until at least 2 entries were added
past the visibly last line, and the JButtons for Previous and Next were
cut off at the top).

Now chunks the user's documents into 500 words each (though it allows for
the last one to be 475, but no less), so the user does not have to worry about chunking. The chunked files are created as temp files in a temp directory, so as far as the user is concerned, the chunking process never happens. These temp files are deleted when the application terminates.

Fixed a bug in the FileHelper read function that added the text "null" to
the beginning of every file read in.